+++ packages/kad-dht/src/query/manager.ts
@@ -143,7 +143,6 @@ export class QueryManager implements Startable {
 
     // query a subset of peers up to `kBucketSize / 2` in length
     let queryFinished = false
-
     try {
       if (this.routingTable.size === 0 && !this.allowQueryWithZeroPeers) {
         log('routing table was empty, waiting for some peers before running%s query', options.isSelfQuery === true ? ' self' : '')
@@ -221,18 +220,7 @@ export class QueryManager implements Startable {
         }
 
         if (event.name === 'PEER_RESPONSE') {
-          for (const peer of [...event.closer, ...event.providers]) {
-            // eslint-disable-next-line max-depth
-            if (!(await this.connectionManager.isDialable(peer.multiaddrs, {
-              signal
-            }))) {
-              continue
-            }
-
-            await this.routingTable.add(peer.id, {
-              signal
-            })
-          }
+          this.routingTable.queueRoutingTableUpdate(event.from)
         }
 
         signal.throwIfAborted()