+++ packages/floodsub/CHANGELOG.md
@@ -98,6 +98,17 @@
   * devDependencies
     * @libp2p/interface-compliance-tests bumped from ^5.3.0 to ^5.3.1
 
+## [11.0.19](https://github.com/libp2p/js-libp2p/compare/floodsub-v11.0.18...floodsub-v11.0.19) (2026-04-25)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @libp2p/interface-internal bumped from ^3.1.2 to ^3.1.3
+    * @libp2p/peer-collections bumped from ^7.0.17 to ^7.0.18
+    * @libp2p/utils bumped from ^7.0.17 to ^7.1.0
+
 ## [11.0.18](https://github.com/libp2p/js-libp2p/compare/floodsub-v11.0.17...floodsub-v11.0.18) (2026-04-16)