+++ packages/peer-collections/CHANGELOG.md
@@ -35,6 +35,15 @@
   * devDependencies
     * @libp2p/peer-id-factory bumped from ^4.0.3 to ^4.0.4
 
+## [7.0.18](https://github.com/libp2p/js-libp2p/compare/peer-collections-v7.0.17...peer-collections-v7.0.18) (2026-04-25)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @libp2p/utils bumped from ^7.0.17 to ^7.1.0
+
 ## [7.0.17](https://github.com/libp2p/js-libp2p/compare/peer-collections-v7.0.16...peer-collections-v7.0.17) (2026-04-16)